% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IObenchmark.R
\name{AD2SCE}
\alias{AD2SCE}
\title{Convert AnnData to SingleCellExperiemnt using scDIOR/zellkonverter/schard.}
\usage{
AD2SCE(
  anndata.file,
  method = c("scDIOR", "zellkonverter", "schard"),
  assay = "RNA",
  slot = "counts",
  use.raw = TRUE,
  conda.path = NULL
)
}
\arguments{
\item{anndata.file}{The file contains AnnData.}

\item{method}{Method used to perform conversion, choose from "scDIOR", "zellkonverter", "schard". Default: "scDIOR".}

\item{assay}{The type of data, used when \code{method} is "scDIOR". Default: "RNA" (scRNA-seq data).}

\item{slot}{Name used when saving adata.X as an assay, used when \code{method} is "zellkonverter". Default: "counts".}

\item{use.raw}{Logical value, whether to use adata.raw. Default: TRUE. When \code{method} is "scDIOR" and \code{use.raw}
is TRUE (FALSE), raw.X -> assays (X/layers -> assays); When \code{method} is "zellkonverter" and \code{use.raw} is TRUE (FALSE),
raw.X -> altExp, X and layers -> assays (X and layers -> assays); When \code{method} is
"schard" and \code{use.raw} is TRUE (FALSE), raw.X -> assays (X -> assays).}

\item{conda.path}{Conda environment path, used when \code{method} is "scDIOR" or "zellkonverter". Default: NULL.}
}
\value{
A SingleCellExperiment object.
}
\description{
Convert AnnData to SingleCellExperiemnt using scDIOR/zellkonverter/schard.
}
\examples{
# scDIOR
sce.scdior <- AD2SCE(anndata.file = "pbmc3k.h5ad", method = "scDIOR", assay = "RNA", use.raw = TRUE, conda.path = "/path/to/conda")
# zellkonverter
sce.zell <- AD2SCE(anndata.file = "pbmc3k.h5ad", method = "zellkonverter", slot = "scale.data", use.raw = TRUE, conda.path = "/path/to/conda")
# schard
sce.schard <- AD2SCE(anndata.file = "pbmc3k.h5ad", method = "schard", use.raw = TRUE)
}
