% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam.R
\name{Bam2Fastq}
\alias{Bam2Fastq}
\title{Convert 10x bam to fastqs.}
\usage{
Bam2Fastq(
  bam.folder = NULL,
  bam.path = NULL,
  bamtofastq.path = NULL,
  bamtofastq.paras = "--nthreads 4"
)
}
\arguments{
\item{bam.folder}{Folder contains bam files, obtained from \code{DownloadSRA}. Default: NULL.}

\item{bam.path}{Paths of bams. \code{bam.folder} and \code{bam.path} cannot be both NULL. Default: NULL.}

\item{bamtofastq.path}{Path to 10x bamtofastq. Default: NULL (conduct automatic detection with bamtofastq_linux).}

\item{bamtofastq.paras}{Parameters for \code{bamtofastq.path}. Default: "--nthreads 4".}
}
\value{
NULL or paths of failed bams.
}
\description{
Convert 10x bam to fastqs.
}
\examples{
# GSE138266.runs = ExtractRun(acce = "GSE138266", platform = "GPL18573")
# GSE138266.down = DownloadBam(gsm.df = GSE138266.runs, prefetch.path = "/path/to/prefetch",
#                              out.folder = "/path/to/output")
# GSE138266.convert = Bam2Fastq(bam.folder = "/path/to/output", bamtofastq.path = "/path/to/bamtofastq_linux",
#                               bamtofastq.paras = "--nthreads 4")
}
