% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIO.R
\name{ExportSeurat}
\alias{ExportSeurat}
\title{Export SeuratObject to Other Formats.}
\usage{
ExportSeurat(
  seu.obj,
  assay = NULL,
  reduction = NULL,
  to = c("SCE", "AnnData", "CellDataSet", "cell_data_set", "loom"),
  anndata.file = NULL,
  loom.file = NULL,
  conda.path = NULL,
  ...
)
}
\arguments{
\item{seu.obj}{A seurat object.}

\item{assay}{Which assay to use. Default: NULL (get with \code{\link{DefaultAssay}}).}

\item{reduction}{Name of DimReduc to set to main reducedDim in cds.}

\item{to}{The target format, chosen from "SCE" (SingleCellExperiment), "AnnData", "CellDataSet", "cell_data_set", "loom".
Default: "SCE".}

\item{anndata.file}{File used to save AnnData results. Default: NULL.}

\item{loom.file}{File used to save loom results. Default: NULL.}

\item{conda.path}{Conda environment path, used when \code{to} is "AnnData". Default: NULL.}

\item{...}{Parameter for \code{\link{as.SingleCellExperiment}}, \code{sceasy::convertFormat}, \code{\link{as.CellDataSet}},
\code{\link{as.cell_data_set}}, \code{\link{SaveLoom}}, corresponding to \code{to}.}
}
\value{
Object corresponding to \code{to}.
}
\description{
Export SeuratObject to Other Formats.
}
