% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSCCellBrowser.R
\name{ExtractCBDatasets}
\alias{ExtractCBDatasets}
\title{Extract Datasets with Attributes.}
\usage{
ExtractCBDatasets(
  all.samples.df,
  collection = NULL,
  sub.collection = NULL,
  organ = NULL,
  disease = NULL,
  organism = NULL,
  project = NULL,
  fuzzy.match = TRUE,
  cell.num = NULL
)
}
\arguments{
\item{all.samples.df}{Dataframe contains all samples metadata, obtained with \code{ShowCBDatasets}.}

\item{collection}{The collection of the datasets, corresponds to \code{shortLabel} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{sub.collection}{The sub-collection of the datasets, corresponds to \code{subLabel} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{organ}{The organ of the datasets, corresponds to \code{body_parts} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{disease}{The disease of the datasets, corresponds to \code{diseases} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{organism}{The specie of the datasets, corresponds to \code{organisms} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{project}{The project of the datasets, corresponds to \code{projects} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{fuzzy.match}{Logical value, whether to perform fuzzy match with provided attribute values. Default: TRUE.}

\item{cell.num}{Cell number filter. If NULL, no filter; if one value, lower filter; if two values, low and high filter. Deault: NULL.}
}
\value{
Dataframe contains filtered datasets.
}
\description{
Extract Datasets with Attributes.
}
\examples{
# # lazy mode, load datasets json files locally
# ucsc.cb.samples = ShowCBDatasets(lazy = TRUE, json.folder = NULL, update = FALSE)
# hbb.sample.df = ExtractCBDatasets(all.samples.df = ucsc.cb.samples, organ = c("brain", "blood"), organism = "Human (H. sapiens)")
}
