% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{ExtractGEOExpSupp}
\alias{ExtractGEOExpSupp}
\title{Extract Raw Count Matrix from Supplementary Files.}
\usage{
ExtractGEOExpSupp(acce, timeout = 3600, supp.idx = 1)
}
\arguments{
\item{acce}{GEO accession number.}

\item{timeout}{Timeout for \code{\link{download.file}}. Default: 3600.}

\item{supp.idx}{The index of supplementary files to download. Default: 1.}
}
\value{
A dataframe.
}
\description{
Extract Raw Count Matrix from Supplementary Files.
}
\examples{
# for bulk rna-seq
# count.mat = ExtractGEOExpSupp(acce = "GSE149838")
# count.mat = ExtractGEOExpSupp(acce = "GSE147507")
# count.mat = ExtractGEOExpSupp(acce = "GSE147507", supp.idx = 2)
# count.mat = ExtractGEOExpSupp(acce = "GSE122774")
# # for single cell matrix
# count.mat = ExtractGEOExpSupp(acce = "GSE94820")
}
