% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{ExtractGEOExpSupp10x}
\alias{ExtractGEOExpSupp10x}
\title{Fortmat Supplementary Files to 10x.}
\usage{
ExtractGEOExpSupp10x(
  acce,
  supp.idx = 1,
  timeout = 3600,
  out.folder = NULL,
  gene2feature = TRUE
)
}
\arguments{
\item{acce}{GEO accession number.}

\item{supp.idx}{The index of supplementary files to download. Default: 1.}

\item{timeout}{Timeout for \code{\link{download.file}}. Default: 3600.}

\item{out.folder}{Output folder to save 10x files. Default: NULL (current working directory).}

\item{gene2feature}{Logical value, whether to rename \code{genes.tsv.gz} to \code{features.tsv.gz}.
Default: TURE.}
}
\description{
Fortmat Supplementary Files to 10x.
}
