% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{ExtractGEOMeta}
\alias{ExtractGEOMeta}
\title{Extract Sample Metadata.}
\usage{
ExtractGEOMeta(pf.obj)
}
\arguments{
\item{pf.obj}{GEO object of platform.}
}
\value{
A dataframe.
}
\description{
Extract Sample Metadata.
}
\examples{
# pf.obj = GEOobj(acce = "GSE94820", platform = "GPL16791")
# pf.info = ExtractGEOMeta(pf.obj)
}
