% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{ExtractGEOSubMeta}
\alias{ExtractGEOSubMeta}
\title{Extract Sample Metadata.}
\usage{
ExtractGEOSubMeta(pf.obj)
}
\arguments{
\item{pf.obj}{GEO object of platform.}
}
\value{
A dataframe.
}
\description{
Extract Sample Metadata.
}
