% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca.R
\name{ExtractHCAMeta}
\alias{ExtractHCAMeta}
\title{Extract Metadata of Human Cell Atlas Projects.}
\usage{
ExtractHCAMeta()
}
\value{
Dataframe contains all available projects.
}
\description{
Extract Metadata of Human Cell Atlas Projects.
}
\examples{
# # all available projects
# all.hca.projects = ExtractHCAMeta()
}
\references{
https://bioconductor.org/packages/release/bioc/html/hca.html
}
