% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq2matrix.R
\name{Fastq2R}
\alias{Fastq2R}
\title{Pipe FASTQ files to SeuratObject and DESeqDataSet.}
\usage{
Fastq2R(
  sample.dir,
  ref,
  method = c("CellRanger", "STAR"),
  localcores = 4,
  localmem = 16,
  thread = 4,
  out.folder = NULL,
  st.path = NULL,
  st.paras = "--chemistry=auto --jobmode=local",
  merge = TRUE,
  count.col = 2,
  meta.data = NULL,
  fmu = NULL
)
}
\arguments{
\item{sample.dir}{Directory contains all samples.}

\item{ref}{Path of folder containing 10x-compatible transcriptome \code{\link{RunCellRanger}}
STAR \code{\link{RunSTAR}} reference.}

\item{method}{Mapping methods, choose from CellRanger (10x Genomics) and STAR (Smart-seq2 or bulk RNA-seq).
Default: CellRanger.}

\item{localcores}{The max cores used \code{\link{RunCellRanger}}. Default: 4.}

\item{localmem}{The max memory (GB) used \code{\link{RunCellRanger}}. Default: 16.}

\item{thread}{The number of threads to use \code{\link{RunSTAR}}. Default: 4.}

\item{out.folder}{Output folder. Default: NULL (current working directory).}

\item{st.path}{Path to \code{STAR} or \code{cellranger}. Default: NULL (conduct automatic detection).}

\item{st.paras}{Parameters for \code{STAR} or \code{cellranger}.
Default: "--chemistry=auto --jobmode=local".}

\item{merge}{Logical, whether to merge the SeuratObjects, use when \code{method} is CellRanger. Default: TRUE.}

\item{count.col}{Column contains used count data (2: unstranded; 3: \code{stranded=yes}; 4: \code{stranded=reverse}),
use when \code{method} is STAR. Default: 2.}

\item{meta.data}{Dataframe contains sample information for DESeqDataSet, use when \code{method} is STAR. Default: NULL.}

\item{fmu}{Column of \code{meta.data} contains group information. Default: NULL.}
}
\value{
SeuratObject, DESeqDataSet or NULL.
}
\description{
Pipe FASTQ files to SeuratObject and DESeqDataSet.
}
\examples{
\dontrun{
# run CellRanger (10x Genomics)
seu <- Fastq2R(
  sample.dir = "/path/to/fastq",
  ref = "/path/to/10x/ref",
  method = "CellRanger",
  out.folder = "/path/to/results",
  st.path = "/path/to/cellranger"
)
# run STAR (Smart-seq2 or bulk RNA-seq)
deobj <- Fastq2R(
  sample.dir = "/path/to/fastq",
  ref = "/path/to/star/ref",
  method = "STAR",
  out.folder = "/path/to/results",
  st.path = "/path/to/STAR",
  st.paras = "--outBAMsortingThreadN 4 --twopassMode None"
)
}
}
