% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PanglaoDBComposition}
\alias{PanglaoDBComposition}
\title{All Sample Composition of PanglaoDB Datasets}
\format{
A data frame with 19,449 rows and 8 variables:
\describe{
  \item{SRA}{The SRA identifier}
  \item{SRS}{The SRS identifier}
  \item{Tissue}{The tissue of the sample}
  \item{Protocol}{The protocol used to generate this sample}
  \item{Species}{The species of this sample}
  \item{Cluster}{Seurat cluster}
  \item{Cells}{Cluster cell number}
  \item{Cell Type}{Predicted cluster cell types}
}
}
\usage{
PanglaoDBComposition
}
\description{
A dataset contains all sample composition of PanglaoDB datasets.
}
\keyword{datasets}
