% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PanglaoDBMeta}
\alias{PanglaoDBMeta}
\title{All Sample Metadata of PanglaoDB Datasets}
\format{
A data frame with 1,368 rows and 7 variables:
\describe{
  \item{SRA}{The SRA identifier}
  \item{SRS}{The SRS identifier}
  \item{Tissue}{The tissue of the sample}
  \item{Protocol}{The protocol used to generate this sample}
  \item{Species}{The species of this sample}
  \item{Cells}{Total cell number of this sample}
  \item{CellType}{Predicted cell types, separated by comma}
}
}
\usage{
PanglaoDBMeta
}
\description{
A dataset contains all sample metadata and cell types of PanglaoDB datasets.
}
\keyword{datasets}
