% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSCCellBrowser.R
\name{ParseCBDatasets}
\alias{ParseCBDatasets}
\title{Download UCSC Cell Browser Datasets.}
\usage{
ParseCBDatasets(
  sample.df = NULL,
  all.samples.df = NULL,
  collection = NULL,
  sub.collection = NULL,
  organ = NULL,
  disease = NULL,
  organism = NULL,
  project = NULL,
  fuzzy.match = TRUE,
  merge = TRUE
)
}
\arguments{
\item{sample.df}{Dataframe contains used datasets. Default: NULL.}

\item{all.samples.df}{Dataframe contains all samples metadata, obtained with \code{ShowCBDatasets}. Default: NULL.
\code{sample.df} and \code{all.samples.df} cannot be both NULL.}

\item{collection}{The collection of the datasets, corresponds to \code{shortLabel} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{sub.collection}{The sub-collection of the datasets, corresponds to \code{subLabel} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{organ}{The organ of the datasets, corresponds to \code{body_parts} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{disease}{The disease of the datasets, corresponds to \code{diseases} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{organism}{The specie of the datasets, corresponds to \code{organisms} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{project}{The project of the datasets, corresponds to \code{projects} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{fuzzy.match}{Logical value, whether to perform fuzzy match with provided attribute values. Default: TRUE.}

\item{merge}{Logical value, whether to merge Seurat list. Default: FALSE.}
}
\value{
Seurat object (if \code{merge} is TRUE) or list of Seurat objects (if \code{merge} is FALSE).
}
\description{
Download UCSC Cell Browser Datasets.
}
\examples{
# # lazy mode, load datasets json files locally
# ucsc.cb.samples = ShowCBDatasets(lazy = TRUE, json.folder = NULL, update = FALSE)
# hbb.sample.seu = ParseCBDatasets(all.samples.df = ucsc.cb.samples, organ = c("brain", "blood"), organism = "Human (H. sapiens)")
}
