% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellxgene.R
\name{ParseCELLxGENE}
\alias{ParseCELLxGENE}
\title{Download CELLxGENE Datasets and Return SeuratObject.}
\usage{
ParseCELLxGENE(
  meta = NULL,
  file.ext = c("rds", "h5ad"),
  out.folder = NULL,
  timeout = 3600,
  quiet = FALSE,
  parallel = TRUE,
  return.seu = FALSE,
  merge = TRUE,
  use.census = FALSE,
  census.version = "stable",
  organism = NULL,
  obs.value.filter = NULL,
  obs.keys = NULL,
  include.genes = NULL,
  ...
)
}
\arguments{
\item{meta}{Metadata used to download, can be from \code{ExtractCELLxGENEMeta},
should contain dataset_id, rds_id/h5ad_id (depend on \code{file.ext}) and name columns.
Skip when \code{use.census} is TRUE. Default: NULL.}

\item{file.ext}{The valid file extension for download. When NULL, use "rds" and "h5ad". Default: c("rds", "h5ad").}

\item{out.folder}{The output folder. Default: NULL (current working directory).}

\item{timeout}{Maximum request time. Default: 3600.}

\item{quiet}{Logical value, whether to show downloading progress. Default: FALSE (show).}

\item{parallel}{Logical value, whether to download parallelly. Default: TRUE. When "libcurl" is available for \code{download.file},
the parallel is done by default (\code{parallel} can be FALSE).}

\item{return.seu}{Logical value, whether to load downloaded datasets to Seurat. Valid when rds in \code{file.ext} and all
datasets download successfully. Default: FALSE.}

\item{merge}{Logical value, whether to merge Seurat list when there are multiple rds files,
used when \code{return.seu} is TRUE. Default: FALSE.}

\item{use.census}{Logical value, whether to use CZ CELLxGENE Census to download and subset datasets. Default: FALSE.}

\item{census.version}{The version of the Census, e.g., "2024-05-13", or "latest" or "stable". Default: stable.}

\item{organism}{Organism, should be in lower case and replace space with '_'. Default: FALSE (human).}

\item{obs.value.filter}{Filter expression for cell's metadata,
e.g., cell_type == 'B cell' & tissue_general == 'lung' & disease == 'COVID-19'. Default: NULL.}

\item{obs.keys}{Columns to fetch for the cell's metadata. e.g., c("cell_type", "tissue_general", "disease", "sex").}

\item{include.genes}{Genes to include, e.g, \code{include.genes} = c('ENSG00000161798', 'ENSG00000188229')
same as \code{var_value_filter} = "feature_id %in% c('ENSG00000161798', 'ENSG00000188229')" in \code{\link{get_seurat}}.
Default: NULL.}

\item{...}{Parameters for \code{\link{get_seurat}}, used when \code{use.census} is TRUE.}
}
\value{
Dataframe contains failed datasets, SeuratObject (\code{return.seu} is TRUE, rds in \code{file.ext}) and
NULL (\code{return.seu} is FALSE or rds not in \code{file.ext}).
}
\description{
Download CELLxGENE Datasets and Return SeuratObject.
}
\examples{
\dontrun{
# all available datasets
all.cellxgene.datasets <- ShowCELLxGENEDatasets()
# human 10x v2 and v3 datasets
human.10x.cellxgene.meta <- ExtractCELLxGENEMeta(
  all.samples.df = all.cellxgene.datasets,
  assay = c("10x 3' v2", "10x 3' v3"),
  organism = "Homo sapiens"
)
# download, need to provide the output folder
ParseCELLxGENE(meta = human.10x.cellxgene.meta, out.folder = "/path/to/output")
}
}
\references{
https://gist.github.com/ivirshup/f1a1603db69de3888eacb4bdb6a9317a
}
