% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{ParseGEO}
\alias{ParseGEO}
\title{Parse GEO Data.}
\usage{
ParseGEO(
  acce,
  platform,
  supp.idx = 1,
  down.supp = FALSE,
  time.out = 3600,
  supp.type = c("count", "10x"),
  out.folder = NULL,
  gene2feature = TRUE,
  ...
)
}
\arguments{
\item{acce}{GEO accession number.}

\item{platform}{Platform information/field.}

\item{supp.idx}{The index of supplementary files to download. This should be consistent with \code{platform}. Default: 1.}

\item{down.supp}{Logical value, whether to download supplementary files to create count matrix. If TRUE, always
download supplementary files. If FALSE, use \code{ExpressionSet} (If contains non-integer or emoty,
download supplementary files automatically). Default: FALSE.}

\item{time.out}{Timeout for \code{\link{download.file}}. Default: 3600.}

\item{supp.type}{The type of downloaded supplementary files, choose from count (count matrix file or single count matrix file)
and 10x (cellranger output files, contains barcodes, genes/features and matrix). Default: count.}

\item{out.folder}{Output folder to save 10x files. Default: NULL (current working directory).}

\item{gene2feature}{Logical value, whether to rename \code{genes.tsv.gz} to \code{features.tsv.gz}. Default: TRUE.}

\item{...}{Parameters for \code{\link{getGEO}}.}
}
\value{
List contains GEO object of platform, study information, raw count matrix and metadata.
}
\description{
Parse GEO Data.
}
\examples{
# # the supp files are count matrix
# GSE94820.list = ParseGEO(acce = "GSE94820", platform = "GPL16791")
# # the supp files are cellranger output files: barcodes, genes/features and matrix
# GSE200257.list = ParseGEO(acce = "GSE200257", platform = "GPL24676", supp.idx = 1, down.supp = TRUE, supp.type = "10x",
#                           out.folder = "/path/to/output/folder")
}
