% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zenodo.R
\name{PrepareZenodo}
\alias{PrepareZenodo}
\title{Prepare Dataframe with Zenodo DOI.}
\usage{
PrepareZenodo(doi, file.ext = c("rdata", "h5ad"))
}
\arguments{
\item{doi}{Zenodo DOI, should start with "10.5281/zenodo.".}

\item{file.ext}{The valid file extension for download. When NULL, use all files. Default: c("rdata", "h5ad").}
}
\value{
Dataframe contains files with valid extension in given Zenodo DOI.
}
\description{
Prepare Dataframe with Zenodo DOI.
}
\examples{
# zebrafish.df = PrepareZenodo(doi = "10.5281/zenodo.7243603")
# PrepareZenodo(doi = "10.5281/zenodo.48065") # Restricted Access
}
