% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq2matrix.R
\name{RunCellRanger}
\alias{RunCellRanger}
\title{Run CellRanger on Downloaded FASTQ Files.}
\usage{
RunCellRanger(
  sample.dir,
  ref,
  localcores = 4,
  localmem = 16,
  out.folder = NULL,
  cr.path = NULL,
  cr.paras = "--chemistry=auto --jobmode=local"
)
}
\arguments{
\item{sample.dir}{Directory contains all samples.}

\item{ref}{Path of folder containing 10x-compatible
transcriptome reference (\code{--transcriptome}).}

\item{localcores}{Set max cores the pipeline may request at one time (\code{--localcores}).
Only applies to local jobs (\code{--jobmode=local}). Default: 4.}

\item{localmem}{Set max GB the pipeline may request at one time (\code{--localmem}).
Only applies to local jobs (\code{--jobmode=local}). Default: 16.}

\item{out.folder}{Output folder. Default: NULL (current working directory).}

\item{cr.path}{Path to cellranger. Default: NULL (conduct automatic detection).}

\item{cr.paras}{Parameters for \code{cellranger}.
Default: "--chemistry=auto --jobmode=local".}
}
\value{
Vector contains failed samples or NULL.
}
\description{
Run CellRanger on Downloaded FASTQ Files.
}
\examples{
\dontrun{
RunCellRanger(
  sample.dir = "/path/to/fastq",
  ref = "/path/to/cellranger_tiny_ref/3.0.0",
  out.folder = "/path/to/results",
  cr.path = "/path/to/cellranger-x.x.x/cellranger"
)
}
}
