% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq2matrix.R
\name{RunSTAR}
\alias{RunSTAR}
\title{Run STAR on Downloaded FASTQ Files.}
\usage{
RunSTAR(
  sample.dir,
  ref,
  out.folder = NULL,
  thread = 4,
  star.path = NULL,
  star.paras = "--outBAMsortingThreadN 4 --twopassMode None"
)
}
\arguments{
\item{sample.dir}{Directory contains all samples.}

\item{ref}{Path of folder containing STAR version-compatible reference.}

\item{out.folder}{Output folder. Default: NULL (current working directory).}

\item{thread}{The number of threads to use. Default: 4.}

\item{star.path}{Path to \code{STAR}. Default: NULL (conduct automatic detection).}

\item{star.paras}{Parameters for \code{STAR}.
Default: "--outBAMsortingThreadN 4 --twopassMode None".}
}
\value{
Vector contains failed samples or NULL.
}
\description{
Run STAR on Downloaded FASTQ Files.
}
\examples{
\dontrun{
RunSTAR(
  sample.dir = "/path/to/fastq",
  ref = "/path/to/star/index",
  out.folder = "/path/to/star/mapping",
  star.path = "/path/to/bin/STAR"
)
}
}
