% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IObenchmark.R
\name{Seu2AD}
\alias{Seu2AD}
\title{Convert SeuratObject to AnnData using SeuratDisk/sceasy/scDIOR.}
\usage{
Seu2AD(
  seu.obj,
  method = c("SeuratDisk", "sceasy", "scDIOR"),
  out.folder = NULL,
  out.filename = NULL,
  assay = "RNA",
  slot = "counts",
  save.scale = FALSE,
  conda.path = NULL
)
}
\arguments{
\item{seu.obj}{A seurat object.}

\item{method}{Method used to perform conversion, choose from "SeuratDisk", "sceasy",    "scDIOR". Default: "SeuratDisk".}

\item{out.folder}{Output folder. Default: NULL (current working directory).}

\item{out.filename}{Output file name, e.g., test.h5ad. Default: NULL (\code{seu.obj} name, method, h5ad.).}

\item{assay}{Assay name to store expression matrices in SeuratObject. Default: RNA.}

\item{slot}{Slot for adata.X, used when \code{method} is "sceasy". Default: counts.}

\item{save.scale}{Logical value, whether to save scaled count matrix, used when \code{method} is "SeuratDisk", "scDIOR".
Default: FALSE. When \code{method} is "SeuratDisk" and \code{save.scale} is TRUE (FALSE), scaled data is stored in X，
log-normalized data is in raw.X (log-normalized data is stored in X，raw count matrix is in raw.X); When \code{method}
is "scDIOR" and \code{save.scale} is TRUE (FALSE), scaled data is stored in X，log-normalized data is in raw.X,
raw count matrix is in layers (log-normalized data is stored in X，raw count matrix is in layers).}

\item{conda.path}{Conda environment path, used when \code{method} is "sceasy". Default: NULL.}
}
\value{
Run log.
}
\description{
Convert SeuratObject to AnnData using SeuratDisk/sceasy/scDIOR.
}
\examples{
# SeuratDisk
Seu2AD(seu.obj = pbmc3k.final, method = "SeuratDisk", out.folder = "benchmark", assay = "RNA", save.scale = TRUE)
# sceasy
Seu2AD(seu.obj = pbmc3k.final, method = "sceasy", out.folder = "benchmark", assay = "RNA", slot = "counts", conda.path = "/path/to/conda")
# scDIOR
Seu2AD(seu.obj = pbmc3k.final, method = "scDIOR", out.folder = "benchmark", assay = "RNA", save.scale = TRUE)
}
