% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSCCellBrowser.R
\name{ShowCBDatasets}
\alias{ShowCBDatasets}
\title{Show All Available Datasets in UCSC Cell Browser.}
\usage{
ShowCBDatasets()
}
\value{
Dataframe contains all available datasets.
}
\description{
Show All Available Datasets in UCSC Cell Browser.
}
\examples{
# ucsc.cb.samples = ShowCBDatasets()
}
