% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellxgene.R
\name{ShowCELLxGENEDatasets}
\alias{ShowCELLxGENEDatasets}
\title{Show All Available Datasets in CELLxGENE.}
\usage{
ShowCELLxGENEDatasets()
}
\value{
Dataframe contains all available datasets.
}
\description{
Show All Available Datasets in CELLxGENE.
}
\examples{
\donttest{
# all available datasets
all.cellxgene.datasets <- ShowCELLxGENEDatasets()
}
}
\references{
https://gist.github.com/ivirshup/f1a1603db69de3888eacb4bdb6a9317a
}
