% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca.R
\name{ShowHCAProjects}
\alias{ShowHCAProjects}
\title{Show All Available Projects in Human Cell Atlas.}
\usage{
ShowHCAProjects()
}
\value{
Dataframe contains all available projects.
}
\description{
Show All Available Projects in Human Cell Atlas.
}
\examples{
# # all available projects
# all.hca.projects = ShowHCAProjects()
}
