% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanglaoDB.R
\name{ShowPanglaoDBMeta}
\alias{ShowPanglaoDBMeta}
\title{Show Metadata of scRNA-seq Datasets in PanglaoDB.}
\usage{
ShowPanglaoDBMeta(
  specie = NULL,
  protocol = NULL,
  tissue = NULL,
  show.cell.type = TRUE
)
}
\arguments{
\item{specie}{The specie of the datasets, choose from "Homo sapiens", "Mus musculus", one or multiple value. Default: NULL (All).}

\item{protocol}{Protocol used to generate the datasets, choose from "10x chromium", "drop-seq", "microwell-seq",
"C1 Fluidigm", "inDrops", "Smart-seq2", "CEL-seq", one or multiple value. Default: NULL (All).}

\item{tissue}{The tissue of the datasets. Default: NULL (All).}

\item{show.cell.type}{Logical value, whether to show inferred cell type. Default: TRUE.}
}
\value{
Dataframe contains SRA, SRS, Tissue, Protocol, Specie, Cells, CellType (inferred).
}
\description{
Show Metadata of scRNA-seq Datasets in PanglaoDB.
}
\examples{
# ShowPanglaoDBMeta(specie = "Homo sapiens", protocol = c("Smart-seq2", "10x chromium"))
}
