% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{StatDBAttribute}
\alias{StatDBAttribute}
\title{Stat Database Attributes.}
\usage{
StatDBAttribute(
  df = NULL,
  filter,
  database = c("PanglaoDB", "UCSC", "CELLxGENE", "HCA"),
  combine = FALSE,
  use.census = FALSE,
  census.version = "stable",
  organism = NULL
)
}
\arguments{
\item{df}{All metadata, can be \code{PanglaoDBMeta} and obtained with \code{ShowCBDatasets}, \code{ShowCELLxGENEDatasets},
and \code{ShowHCAProjects}. Skip when \code{use.census} is TRUE. Default: NULL.}

\item{filter}{Vector of attributes.}

\item{database}{Database name, choose from "PanglaoDB", "UCSC", "CELLxGENE", "HCA". Default: "PanglaoDB".}

\item{combine}{Logical value, whether to combine all attributes in \code{filter} for summarising.
Default: FALSE.}

\item{use.census}{Logical value, whether to use CZ CELLxGENE Census to summarise metadata. Default: FALSE.}

\item{census.version}{The version of the Census, e.g., "2024-05-13", or "latest" or "stable". Default: stable.}

\item{organism}{Organism, should be in lower case and replace space with '_'. Default: FALSE (human).}
}
\value{
List of attributes information (attribute, value and number) when \code{combine} is FALSE,
dataframe when \code{combine} is TRUE.
}
\description{
Stat Database Attributes.
}
\examples{
\dontrun{
# PanglaoDB
StatDBAttribute(df = PanglaoDBMeta, filter = c("species", "protocol"), database = "PanglaoDB")
# UCSC Cell Browser, need users to provide the json folder
ucsc.cb.samples <- ShowCBDatasets(lazy = TRUE, json.folder = NULL, update = FALSE)
StatDBAttribute(df = ucsc.cb.samples, filter = c("organism", "organ"), database = "UCSC")
# CELLxGENE
all.cellxgene.datasets <- ShowCELLxGENEDatasets()
StatDBAttribute(
  df = all.cellxgene.datasets, filter = c("organism", "sex"),
  database = "CELLxGENE"
)
# HCA
all.hca.projects <- ShowHCAProjects()
StatDBAttribute(df = all.hca.projects, filter = c("organism", "sex"), database = "HCA")
}
}
