% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam.R
\name{DownloadBam}
\alias{DownloadBam}
\title{Download bam.}
\usage{
DownloadBam(
  gsm.df,
  prefetch.path = NULL,
  out.folder = NULL,
  prefetch.paras = "-X 100G"
)
}
\arguments{
\item{gsm.df}{Dataframe contains GSM and Run numbers, obtained from \code{ExtractRun}.}

\item{prefetch.path}{Path to prefetch. Default: NULL (conduct automatic detection).}

\item{out.folder}{Output folder. Default: NULL (current working directory).}

\item{prefetch.paras}{Parameters for \code{prefetch}. This should not contain --type or -T values. Default: "-X 100G".}
}
\value{
Dataframe contains failed runs or NULL.
}
\description{
Download bam.
}
\examples{
# GSE138266.runs = ExtractRun(acce = "GSE138266", platform = "GPL18573")
# GSE138266.down = DownloadBam(gsm.df = GSE138266.runs, prefetch.path = "/path/to/prefetch",
#                              out.folder = "/path/to/output")
}
