% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSCCellBrowser.R
\name{ExtractCBDatasets}
\alias{ExtractCBDatasets}
\title{Extract Datasets with Attributes.}
\usage{
ExtractCBDatasets(
  collection = NULL,
  sub.collection = NULL,
  organ = NULL,
  disease = NULL,
  organism = NULL,
  project = NULL,
  fuzzy.match = TRUE
)
}
\arguments{
\item{collection}{The collection of the datasets, corresponds to \code{shortLabel} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{sub.collection}{The sub-collection of the datasets, corresponds to \code{subLabel} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{organ}{The organ of the datasets, corresponds to \code{body_parts} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{disease}{The disease of the datasets, corresponds to \code{diseases} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{organism}{The specie of the datasets, corresponds to \code{organisms} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{project}{The project of the datasets, corresponds to \code{projects} column
of \code{ShowCBDatasets}. Default: NULL (without filtering).}

\item{fuzzy.match}{Logical value, whether to perform fuzzy match with provided attribute values. Default: TRUE.}
}
\value{
Dataframe contains filtered datasets.
}
\description{
Extract Datasets with Attributes.
}
\examples{
#
}
