% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellxgene.R
\name{ExtractCELLxGENEMeta}
\alias{ExtractCELLxGENEMeta}
\title{Extract Metadata of CELLxGENE Datasets.}
\usage{
ExtractCELLxGENEMeta()
}
\value{
Dataframe contains all available datasets.
}
\description{
Extract Metadata of CELLxGENE Datasets.
}
\examples{
# # all available datasets
# cellxgene.meta = ExtractCELLxGENEMeta()
}
