% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca.R
\name{ShowHCAProjects}
\alias{ShowHCAProjects}
\title{Show All Available Projects in Human Cell Atlas.}
\usage{
ShowHCAProjects(catalog = NULL)
}
\arguments{
\item{catalog}{The catalog of the projects. Different catalogs may share some projects. Choose from "dcp29",
"dcp30", "dcp1", "lm2", "lm3", one or multiple values. Default: NULL (all catalogs, remove duplicated projects).}
}
\value{
Dataframe contains all available projects.
}
\description{
Show All Available Projects in Human Cell Atlas.
}
\examples{
\donttest{
# all available projects
all.hca.projects <- ShowHCAProjects()
}
}
